:white_check_mark: Fix CI workflow on public repository
The github secret token is used to run a Github Actions workflow file
on a public repository. It is important here that we have branch
protection rules that do not allow improper use of the actions, for e.g:
- printing out repository/organization secrets
- infinite workflow runs, etc.

It is advised to have branch protection rules such as requiring an
approval before running workflows from new users, and approvals to merge
such PRs.